Move `ShapeProcessor` to `fj-operations`
It can't stay in `fj-app`, as the export code (which I intend to move
into its own crate) depends on it. I think `fj-operations` is a good
place for it, as the code that's already there has practically the same
dependencies as `ShapeProcessor`.
